Refactoring project load classes
