+++ pulse/project/project.py
@@ -27,7 +27,6 @@ class Project:
     def __init__(self):
 
         self.file = app().file
-        self.loader = LoadProjectData()
         self.preprocessor = Preprocessor(self)        
 
         self.reset()
@@ -388,174 +387,8 @@ class Project:
         return Entity(tag)
 
     def load_entity_file(self):
-        try:
-
-            self.loader.load_project_data_from_files()
-            self.number_sections_by_line = dict()
-            # self.lines_with_cross_section_by_elements = list()
-
-            # Load structural element type info
-            for key, etype_data in self.loader.structural_element_type_data.items():
-                if self.loader.element_type_is_structural:
-                    if "-" in key:
-                        self.load_structural_element_type_by_elements(etype_data[0], 
-                                                                      etype_data[1])
-                    else:
-                        line_id = int(key)
-                        self.load_structural_element_type_by_line(line_id, 
-                                                                  etype_data) 
-
-            # Structural element force offset to the entities
-            for key, force_offset_data in self.loader.structural_element_force_offset_data.items():
-                if "-" in key:
-                    self.load_structural_element_force_offset_by_elements(force_offset_data[0], 
-                                                                          force_offset_data[1])
-                else:
-                    line_id = int(key)
-                    self.load_structural_element_force_offset_by_line(line_id, 
-                                                                      force_offset_data)
-
-            # Structural element wall formulation to the entities
-            for key, wall_formulation_data in self.loader.structural_element_wall_formulation_data.items():
-                if "-" in key:
-                    self.load_structural_element_wall_formulation_by_elements(wall_formulation_data[0], 
-                                                                              wall_formulation_data[1])
-                else:
-                    line_id = int(key)
-                    self.load_structural_element_wall_formulation_by_line(line_id, 
-                                                                          wall_formulation_data)
-
-            # Acoustic element type to the entities
-            for key, [el_type, proportional_damping, vol_flow] in self.loader.acoustic_element_type_data.items():
-                if self.loader.element_type_is_acoustic:
-                    self.load_acoustic_element_type_by_line(key, el_type, proportional_damping=proportional_damping, vol_flow = vol_flow)
-
-            # Length correction to the elements
-            for key, value in self.loader.element_length_correction_data.items():
-                self.load_length_correction_by_elements(value[0], value[1], key)
-
-            for key, value in self.loader.perforated_plate_data.items():
-                frequency_setup_pass = True
-                table_name = value[1].dimensionless_impedance_table_name
-                if table_name is not None:
-                    if self.change_project_frequency_setup(table_name, value[2]):
-                        frequency_setup_pass = False
-                        break
-                if frequency_setup_pass:
-                    self.load_perforated_plate_by_elements(value[0], value[1], key)
-
-            # Material to the entities
-            for key, mat in self.loader.material_data.items():
-                self.load_material_by_line(key, mat)
-
-            # Fluid to the entities
-            for key, fld in self.loader.fluid_data.items():
-                self.load_fluid_by_line(key, fld)
-
-            # Compressor info to the entities
-            for key, _info in self.loader.compressor_info.items():
-                self.load_compressor_info_by_line(key, _info)
-
-            # Constant cross-section to the entities
-            for key, section_data in self.loader.cross_section_data.items():
-                # key[0] = tag : str
-                # key[1] = label ("pipe", "beam")
-                if "-" in key[0]:
-                    
-                    cross = CrossSection(pipe_section_info = section_data[1])
-                    self.load_cross_section_by_element(section_data[0], cross)
-                    
-                    prefix_key = int(key[0].split("-")[0])
-                    if prefix_key in list(self.number_sections_by_line.keys()):
-                        self.number_sections_by_line[prefix_key] += 1
-                    
-                    else:
-                        self.number_sections_by_line[prefix_key] = 1
-
-                else:
-
-                    if key not in self.loader.variable_sections_data.keys():
-
-                        cross = None
-                        if key[1] == "pipe":
-                            cross = CrossSection(pipe_section_info = section_data)
-                        
-                        elif key[1] == "beam":
-                            cross = CrossSection(beam_section_info = section_data)
-                        
-                        if cross is not None:
-                            self.load_cross_section_by_line(int(key[0]), cross)
-            
-            # Variable Cross-section to the entities
-            for key, value in self.loader.variable_sections_data.items():
-                self.load_variable_cross_section_by_line(int(key[0]), value)
-
-            # Beam X-axis rotation to the entities
-            for key, angle in self.loader.beam_xaxis_rotation_data.items():
-                self.load_beam_xaxis_rotation_by_line(key, angle)
-            if len(self.loader.beam_xaxis_rotation_data) > 0:
-                self.preprocessor.process_all_rotation_matrices() 
-
-            # B2PX Rotation Decoupling
-            for key, item in self.loader.B2XP_rotation_decoupling_data.items():
-                if "B2PX ROTATION DECOUPLING" in str(key):
-                    self.preprocessor.dict_B2PX_rotation_decoupling[str(item[2])] = [item[0], item[1], key]
-                    for i in range(len(item[0])):
-                        self.load_B2PX_rotation_decoupling(item[0][i], item[1][i], rotations_to_decouple=item[2])
-
-            # Expansion Joint to the entities
-            for key, joint_data in self.loader.expansion_joint_parameters_data.items():
-                frequency_setup_pass = True
-                if "-" in key:
-                    parameters = joint_data[1]
-                    joint_tables = joint_data[1][2]
-                    joint_list_freq = joint_data[1][3]
-                    if joint_data[1][1] is not None:
-                        for i, joint_freq in enumerate(joint_list_freq):
-                            if self.change_project_frequency_setup(joint_tables[i], joint_freq):
-                                frequency_setup_pass = False
-                                break
-                    if frequency_setup_pass:
-                        for list_elements in joint_data[0]:
-                            self.load_expansion_joint_by_elements(list_elements, parameters[:-1])
-                else:
-                    joint_table_names = joint_data[2]
-                    joint_list_freq = joint_data[3]
-                    if joint_data[1] is not None:
-                        for i, joint_freq in enumerate(joint_list_freq):
-                            if self.change_project_frequency_setup(joint_table_names[i], joint_freq):
-                                frequency_setup_pass = False
-                                break
-                        if frequency_setup_pass:
-                            line_id = int(key)
-                            self.load_expansion_joint_by_lines(line_id, joint_data[:-1])
-            
-            # Valve to the entities
-            for key, [valve_data, cross_sections] in self.loader.valve_data.items():
-                if "-" in key:
-                    self.load_valve_by_elements(valve_data, cross_sections)                      
-                else:
-                    line_id = int(key)
-                    self.load_valve_by_lines(line_id, valve_data, cross_sections)            
-                
-            # Stress Stiffening to the entities and elements
-            for key, parameters in self.loader.stress_stiffening_data.items():
-                if "STRESS STIFFENING" in str(key):
-                    self.load_stress_stiffening_by_elements(parameters[0], parameters[1], section=key)
-                else:
-                    self.load_stress_stiffening_by_line(key, parameters)        
-
-            # Capped end to the entities and elements
-            for key, group in self.loader.capped_end_data.items():
-                if "CAPPED END" in key:  
-                    self.load_capped_end_by_elements(group, True, key)
-                else:
-                    self.load_capped_end_by_line(group, True)
-
-        except Exception as log_error:
-            title = "Error reached while loading data from dictionaries"
-            message = str(log_error)
-            PrintMessageInput([window_title, title, message])
+        self.loader = LoadProjectData()
+        self.loader.load_entity_file()
 
     def process_cross_sections_mapping(self):
